Factored out expandable-on-overflow logic
