+++ plugin-hrm-form/src/components/search/ExpandableTextBlock.tsx
@@ -1,8 +1,8 @@
-import React, { useEffect, useRef } from 'react';
+import React from 'react';
 import type { StyledProps } from '@material-ui/core';
 import { Template } from '@twilio/flex-ui';
 
-import { useIsOverflowing } from './detectOverflow';
+import { useExpandableOnOverflow } from '../../hooks/useExpandableOnOverflow';
 import { StyledLink } from '../../styles/search';
 
 export type ExpandableTextBlockProps = {
@@ -16,30 +16,22 @@ const ExpandableTextBlock: React.FC<ExpandableTextBlockProps & Partial<StyledPro
   collapseLinkText,
   className,
 }) => {
-  const ref = React.useRef();
-  const isOverflowing = useIsOverflowing(ref);
-  const [isExpanded, setExpanded] = React.useState(false);
-  const expandButtonElement = useRef<HTMLButtonElement>(undefined);
-  const collapseButtonElement = useRef<HTMLButtonElement>(undefined);
-
-  const handleExpand = () => {
-    setExpanded(true);
-  };
-
-  const handleCollapse = () => {
-    setExpanded(false);
-  };
-
-  useEffect(() => {
-    if (isExpanded) {
-      collapseButtonElement.current?.focus();
-    } else {
-      expandButtonElement.current?.focus();
-    }
-  }, [isExpanded]);
+  const {
+    collapseButtonElementRef,
+    expandButtonElementRef,
+    handleCollapse,
+    handleExpand,
+    isExpanded,
+    isOverflowing,
+    overflowingRef,
+  } = useExpandableOnOverflow({});
 
   return (
-    <div className={className} style={{ display: 'flex', flexFlow: 'row', justifyContent: 'stretch' }} ref={ref}>
+    <div
+      className={className}
+      style={{ display: 'flex', flexFlow: 'row', justifyContent: 'stretch' }}
+      ref={overflowingRef}
+    >
       <div
         style={{
           textOverflow: 'ellipsis',
@@ -54,14 +46,14 @@ const ExpandableTextBlock: React.FC<ExpandableTextBlockProps & Partial<StyledPro
           underline={true}
           type="button"
           onClick={handleCollapse}
-          ref={collapseButtonElement}
+          ref={collapseButtonElementRef}
           style={{ display: isExpanded ? 'inline' : 'none', marginTop: -3.5 }}
         >
           <Template code={collapseLinkText} />
         </StyledLink>
       </div>
       <div style={{ whiteSpace: 'nowrap', display: isOverflowing && !isExpanded ? 'inherit' : 'none' }}>
-        <StyledLink underline={true} onClick={handleExpand} ref={expandButtonElement} style={{ marginTop: -3.5 }}>
+        <StyledLink underline={true} onClick={handleExpand} ref={expandButtonElementRef} style={{ marginTop: -3.5 }}>
           <Template code={expandLinkText} />
         </StyledLink>
       </div>
